Number of moves semi-implimented
Game Manager is now counting the number of tiles moved to calculate action points. If the player exceeds the number of acton points, the does not allow them to move.

Also Added multi tile hazards, starting with a 2x1 hazard tile.

Need to have the child of the hazard also check if the player is stainding on top of them.